feat: expand packaging metadata
